fix: resolve floating-point legging bugs, integer truncation, and test mode bugs
- Fixed floating-point precision legging risk during atomic MarketBuyArb
- Removed integer truncation for trade size compounding calculations
- Fixed allowance validation checks for test mode in clob client
- Updated liquidity tests to correctly validate against 100% safety margin
- Permanently disabled TestMode flag in config

+++ cmd/paperbot/main.go
@@ -1217,7 +1217,7 @@ func runTrader(ctx context.Context, t *MarketTrader) (*marketResult, error) {
 
 						// Apply compounding multiplier from profitable rounds
 						compoundMult := t.Engine.GetCompoundMultiplier()
-						shares = float64(int(float64(shares) * compoundMult))
+						shares = shares * compoundMult
 
 						// Force at least 1 share if there's any matched liquidity and we have budget
 						if shares < 1.0 && minLiquidity >= 1.0 {
